apps/wedding-platform/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
redis==5.0.8
django-redis==5.4.0
Pillow==10.4.0
boto3==1.35.0
django-storages==1.14.4
python-decouple==3.8
gunicorn==22.0.0
pytest-django==4.8.0
pytest==8.3.2
factory-boy==3.3.1
